# Conflicts:
#	vars/rpr_blender28plugin_pipeline.groovy

+++ vars/rpr_viewer_pipeline.groovy
@@ -232,17 +232,19 @@ def executeTests(String osName, String asicName, Map options)
 
                     // reallocate node if there are still attempts
                     if (sessionReport.summary.total == sessionReport.summary.error + sessionReport.summary.skipped) {
-                        collectCrashInfo(osName, options)
-                        if (osName == "Ubuntu18"){
-                            sh """
-                                echo "Restarting Unix Machine...."
-                                hostname
-                                (sleep 3; sudo shutdown -r now) &
-                            """
-                            sleep(60)
-                        }
-                        if (options.currentTry < options.nodeReallocateTries) {
-                            throw new Exception("All tests crashed")
+                        if (sessionReport.summary.total != sessionReport.summary.skipped){
+                            collectCrashInfo(osName, options)
+                            if (osName == "Ubuntu18"){
+                                sh """
+                                    echo "Restarting Unix Machine...."
+                                    hostname
+                                    (sleep 3; sudo shutdown -r now) &
+                                """
+                                sleep(60)
+                            }
+                            if (options.currentTry < options.nodeReallocateTries) {
+                                throw new Exception("All tests crashed")
+                            }
                         }
                     }
                 }
